Squashed commit of the following:

commit a45424a74c8ebcef1862fc8ad2fcebc436c67c84
Author: Michael Altizer <mialtize@cisco.com>
Date:   Thu Jun 2 16:03:35 2016 -0400

    connectors: Add CppUTest CPPFLAGS to file_connector_module_test build

commit 7304302f6ba0ff6dd52343b67e17b13a2f7696f1
Author: Michael Altizer <mialtize@cisco.com>
Date:   Thu Jun 2 15:50:51 2016 -0400

    debug: Enable argument checking for DebugFormat and fix most fallout

commit 3f67a139df7978e00c7901e88a58ef19b05d1380
Author: Michael Altizer <mialtize@cisco.com>
Date:   Wed Jun 1 13:14:52 2016 -0400

    unprivileged: Initial privilege dropping support

commit be6e4d1591a323e5cf59c4666a6869d08dfbd84c
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue May 31 15:56:16 2016 -0400

    extra: Fix CPPFLAGS variable expansion and add m4 directory

commit 1985958afa96b634dc9a7cdc312a21c61e2774de
Author: Michael Altizer <mialtize@cisco.com>
Date:   Thu May 19 12:13:20 2016 -0400

    sfdaq: Remove reference to daq-mode option

commit 92efb5a5a3bfba1c9a8c8074d1f150d51ceb7687
Author: Michael Altizer <mialtize@cisco.com>
Date:   Wed May 18 12:48:29 2016 -0400

    packet_capture: Fix compiler warnings

commit f182c36194e63a0871e9c42f233b8dbc05a7666a
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue May 10 18:21:47 2016 -0400

    Remove unnecessary utils/stats.h includes

commit 5b278b811793511b732d6549eb447429968d29e5
Author: Michael Altizer <mialtize@cisco.com>
Date:   Mon May 9 12:32:12 2016 -0400

    Remove unnecessary analyzer.h includes

+++ src/ips_options/ips_ipopts.cc
@@ -134,7 +134,7 @@ int IpOptOption::eval(Cursor&, Packet* p)
     for ( const ip::IpOptions& opt : iter)
     {
         DebugFormat(DEBUG_IPS_OPTION, "testing pkt(%d):rule(%d)\n",
-            ipOptionData->ip_option,
+            static_cast<int>(ipOptionData->ip_option),
             static_cast<int>(opt.code));
 
         if (ipOptionData->ip_option == opt.code)
